Moved function plugin into own package
