+++ karma.conf.js
@@ -20,11 +20,15 @@ module.exports = config => {
       'node_modules/es5-shim/es5-sham.js',
       'node_modules/css.escape/css.escape.js',
       './packages/jss/tests/index.js',
-      './packages/jss-observable/src/observable.test.js'
+      './packages/jss-observable/src/observable.test.js',
+      './packages/jss-plugin-syntax-rule-value-function/src/function-rules.test.js',
+      './packages/jss-plugin-syntax-rule-value-function/src/function-values.test.js'
     ],
     preprocessors: {
       './packages/jss/tests/index.js': ['webpack', 'sourcemap'],
-      './packages/jss-observable/src/observable.test.js': ['webpack', 'sourcemap']
+      './packages/jss-observable/src/observable.test.js': ['webpack', 'sourcemap'],
+      './packages/jss-plugin-syntax-rule-value-function/src/function-rules.test.js': ['webpack', 'sourcemap'],
+      './packages/jss-plugin-syntax-rule-value-function/src/function-values.test.js': ['webpack', 'sourcemap']
     },
     webpack: Object.assign(webpackConfig, {
       devtool: 'inline-source-map'
