+++ packages/jss/tests/index.js
@@ -17,8 +17,6 @@ import './integration/rules'
 import './integration/sheet'
 import './integration/sheetsRegistry'
 import './integration/plugins'
-import './functional/function-values'
-import './functional/function-rules'
 import './functional/rules'
 import './functional/sheet'
 import './functional/priority'
